Implementa ingestao via agente, analytics (cache/forecast), PDVs/campanhas e ajustes de deploy/docs

+++ backend/src/main/java/com/pdv2cloud/model/dto/ProductAnalyticsDTO.java
@@ -14,5 +14,5 @@ public class ProductAnalyticsDTO {
     private BigDecimal revenue;
     private BigDecimal quantitySold;
     private BigDecimal averagePrice;
-    private int transactionCount;
+    private long transactionCount;
 }
